fix: column alignment of popup
This commit fixes the issue of column alignment with the cursor.
It makes it so that the popup will always be shown from the start
of the word that is renamed, similar to how VS Code does it.

Besides it also starts work on showing the popup above the current
row, if there is not enough space for it below the current line. It
also starts taking into account the border when computing the line
at which the popup should be displayed.