feat: Rename on submit
This commit finally add the rename on submit feature, which makes the
basis of this plugin.

Besides this functionality, it also makes better use of
`plenary.popup`'s API, leaving it to highlight the window and its
border with the proper colors, instead of manually setting them in the
plugin.

Finally, in this commit some refactoring is done to prepare for an
attempt to support "prompt"-like behaviour of the `renamer.nvim` popup.