test: Add setup tests
This commit adds the initial unit tests (for the `setup` method) as well
as the required files for the GitHub Actions workflow which should
always validate tests, for pushes and pull requests.

It also adds a Makefile for easy test running and setups StyLua for
style checking (in the same workflow used to run the tests).