test: Add `close` spec and refactor
This commit adds the `close` spec for testing the `on_submit` and
`on_close` behaviours. It also refactors existing code to follow the
`StyLua` guidelines.

While not all of the added tests are currently working as intended, this
commit is aimed at synchronizing the local changes with the remote ones.
Thus, further commits will address the failing tests and expand on the
current ones.

Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>