ci: Add `latest` build
Currently, the only target tests are run on is the `nightly` build.
However, since it might fail and might not be the one used by most
users, it is mandatory to also check the latest release version.

The simple fix is to make a `curl` request to GitHub to get the version
tag and then continue with the setup in the same manner as with the
`nightly` target, which is done in this commit.

Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>