ci: fix workflow triggers
Workflows (GitHub Actions) were run on a push basis, regardless of
whether or not a pull request was made for that branch. In order to
preserve bandwidth and be more efficient, this commit changes the
triggers to only run workflows for pushes and pull requests on the
`master`, `develop` and `release` branches.

Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>