docs(neovim): add Neovim documentation
In order to make it easy to reference the plugin API and documentation
inside Neovim, the project should define a `doc/renamer.txt` file. The
contents of the file should be in the Vim "help-file" syntax and have
the specific footer of such files.

The documentation file created in this commit contains an introduction
with a kind of "Table of contents" for the two public functions, `setup`
and `rename`.

Related issue: GH-7
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>