ci: fix triggers
Triggers for `.github/workflows/pr_labeler.yaml` and
`.github/workflows/release.yaml` do not work as intended. For the
former, labeling is triggered for pushes, not only pull requests,
leading to failures. For the latter, the branch name should be `master`,
as the repository does not define a `main` branch.

Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>