ci: fix ci workflows
Triggers for `.github/workflows/pr_labeler.yaml` and
`.github/workflows/release.yaml` do not work as intended. For the
former, labeling is triggered for pushes, not only pull requests,
leading to failures. For the latter, the branch name should be `master`,
as the repository does not define a `main` branch.

In order to follow the
[Gitflow Workflow](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow)
conventions, the `hotfix` label should also be included in the branch
based labels, which is done in this commit. The step responsible for
this also receives a friendlier name in this commit.

Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>