docs(readme.md): add `master` build badge
Add a build status badge for the `master` branch workflow runs, in order
to show the status of the latest build. This is needed as users *should*
use this branch when installing and updating the plugin. Hence, the need
to easily show them if there are any issues with the latest version.

To make it possible with GitHub Actions, a copy of the existing
`.github/workflow/ci.yaml` workflow needs to be made. The new copy will
only run on changes on the `master` branch, while the existing only on
the `develop` and `release/**` branches.

Resolves: GH-22
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>