fix(keymaps): `<c-c>` added an `i` after clear
Since `<c-c>` uses the `0C` normal mode keymaps behind the scenes and
`C` puts the user in `insert` mode, there is no need to re-enter it.

The fix in this case is not to set `keep_insert` to `true`, in the
`exec_in_normal()` parameters, from
`require('renamer.mappings.utils').clear_line()`.

Closes: GH-30
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>