fix: cursor placement when finishing `rename()`
cdb40c6 (GH-33) introduced a bug which made the cursor go 1 character
back when finishing the `rename()` call (by cancellation or submission).

To fix this, the popup options also contain various information
regarding the starting position (where the word starts and the cursor
position at the moment of the call). This makes is easy to update
`on_close()` to set the cursor at the initial position on cancellation
(behaviour toggled by a new parameter).

This commit also makes it so as, upon submission of the rename, the
cursor is placed at the end of the new word, by copying the current
implementation of `vim.lsp.buf.rename()`. The new `_lsp_rename` method
makes LSP requests in the same way Neovim implements the rename feature
and uses a callback to asynchronously set the cursor position.

Fixes: GH-34
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>