fix: cursor position when cancelled from `insert`
In `insert` mode, when cancelling `rename()` was cancelled, the cursor
is put one character ahead of its initial position, as a result of
cdcdb17. This commit fixes the issue by checking for the initial mode in
which `rename()` was called, for `on_close()` or by getting the current
mode and checking it, for `on_submit()`.

Fixes: GH-37
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>