docs(highlights): add vimdocs for highlighting
Update the Vim help-file (`doc/renamer.txt`) to contain a section on
highlights. This is needed as there is currently no way to find out what
highlight groups are defined by the plugin (or which are customizable),
while inside of Neovim. One might not even ever know there are any, if
they would not look in the `README.md`, rather only in the vimdocs.

Closes: GH-46
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>