fix!: popup draws correctly on screen edges
Offset column position to the left, when the cursor is at the end of the
window. This affects the popup with and without border.

BREAKING CHANGE: If the `cword` has a length greater than the `width`,
set `width` to that of the `cword`. Add default width of 0, if `title`
is not set.

Closes: GH-5
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>