feat!: add `with_popup` to enable/disable popup
Add the `with_popup` parameter to determine whether or not to use the
popup to enter the new name. If popup is disabled, use `vim.fn.input`
to receive the new name.

BREAKING CHANGE: Fallback when popup is disabled/not able to be drawn
changed from `vim.lsp.buf.rename()` to using `vim.fn.input()` to ask the
user for input. This is required for:

- setting the quickfix list with the resulting changes
- preserving non-GUI input alternative, since Neovim 0.5.2+ will have
  its own interface for `vim.lsp.buf.rename()`

Closes: GH-68
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>