fix: popup having `sidescrolloff` if set
The popup would have the `sidescrolloff` set in Neovim's config, leading
to words being truncated and looking off. Now the setting is put to 0
for the popup window to prevent the issue.

There was also an issue with the width of the popup if the word to be
renamed was larger than the default width, which lead to 1 character
being truncated. This was fixed by adding `+ 1` to the word length.

Closes: GH-101
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>